+++ usr/src/cmd/fm/modules/common/eversholt/config_impl.h
@@ -19,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2008 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -43,7 +43,7 @@ struct config {
 	struct config *child;
 	struct config *parent;
 	const char *s;
-	int num;
+	intptr_t num;
 	struct lut *props;
 };
